ai-ops/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.1
PyJWT[crypto]==2.8.0 